requirements.txt
asgiref==3.5.0
astroid==2.9.3
colorama==0.4.4
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.5.1
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.2
pylint-plugin-utils==0.7
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
wrapt==1.13.3
asgiref==3.5.0
astroid==2.9.3
colorama==0.4.4
Django==4.0.3
django-ckeditor==6.2.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.5.1
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.2
pylint-plugin-utils==0.7
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
wrapt==1.13.3
